# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/ottosfa/v2/merchant-new-recruitment/create": {
            "post": {
                "description": "Merchant New Recruitment - Create",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Merchant New Rec"
                ],
                "summary": "Merchant New Recruitment - Create",
                "operationId": "Merchant New Recruitment - Create",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 7n1cdTNMK_EeoXr_0D0luYJ68NaYESWxyPXd",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dbmodels.MerchantNewRecruitments"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Merchant New Recruitment - Create EXAMPLE",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/ottosfa/v2/merchant-new-recruitment/download-template": {
            "get": {
                "description": "Merchant New Recruitment - Download Template",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Merchant New Rec"
                ],
                "summary": "Merchant New Recruitment - Download Template",
                "operationId": "Merchant New Recruitment - Download Template",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 7n1cdTNMK_EeoXr_0D0luYJ68NaYESWxyPXd",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Merchant New Recruitment - Download Template EXAMPLE",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/ottosfa/v2/merchant-new-recruitment/export": {
            "post": {
                "description": "Merchant New Recruitment - Export",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Merchant New Rec"
                ],
                "summary": "Merchant New Recruitment - Export",
                "operationId": "Merchant New Recruitment - Export",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 7n1cdTNMK_EeoXr_0D0luYJ68NaYESWxyPXd",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MerchantNewRecruitmentListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Merchant New Recruitment - Export EXAMPLE",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/ottosfa/v2/sub-area-channel/list": {
            "post": {
                "description": "SubAreaChannel - List",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "List"
                ],
                "summary": "SubAreaChannel - List",
                "operationId": "SubAreaChannel - List",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SearchReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sub Area Channel - List Response EXAMPLE",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dbmodels.SubArea"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/ottosfa/v2/todolist/download-template": {
            "get": {
                "description": "Todolist - Download Template",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Todolist"
                ],
                "summary": "Todolist - Download Template",
                "operationId": "Todolist - Download Template",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 7n1cdTNMK_EeoXr_0D0luYJ68NaYESWxyPXd",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Todolist - Download Template EXAMPLE",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dbmodels.MerchantNewRecruitments": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "Jalan Buntu no 21"
                },
                "city_id": {
                    "type": "integer",
                    "example": 3171
                },
                "created_at": {
                    "type": "string"
                },
                "customer_code": {
                    "type": "string",
                    "example": "1234566543216.0"
                },
                "district_id": {
                    "type": "integer",
                    "example": 3171100
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "id_card": {
                    "type": "string",
                    "example": "12345678909887654"
                },
                "institution_code": {
                    "type": "string",
                    "example": "OP"
                },
                "latitude": {
                    "type": "string",
                    "example": "-6.2354219"
                },
                "longitude": {
                    "type": "string",
                    "example": "106.81721122"
                },
                "name": {
                    "type": "string",
                    "example": "Test Data 2"
                },
                "owner_name": {
                    "type": "string",
                    "example": "Namaku Siapa Ya"
                },
                "phone_number": {
                    "type": "string",
                    "example": "088999121367"
                },
                "province_id": {
                    "type": "integer",
                    "example": 31
                },
                "sales_type_id": {
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "type": "string",
                    "example": "Activated"
                },
                "sub_area_channel_id": {
                    "type": "integer",
                    "example": 4
                },
                "sub_area_channel_name": {
                    "type": "string",
                    "example": "Sub Area Jabar 0004"
                },
                "updated_at": {
                    "type": "string"
                },
                "village_id": {
                    "type": "integer",
                    "example": 3171100008
                }
            }
        },
...
        "dbmodels.SubArea": {
            "type": "object",
            "properties": {
                "area_id": {
                    "type": "integer"
                },
                "code": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "models.MerchantNewRecruitmentListReq": {
            "type": "object",
            "properties": {
                "customer_code": {
                    "type": "string",
                    "example": "566543"
                },
                "id": {
                    "type": "string",
                    "example": "2"
                },
                "institution_code": {
                    "type": "string",
                    "example": "OP"
                },
                "name": {
                    "type": "string",
                    "example": "Data"
                },
                "page": {
                    "type": "integer"
                },
                "phone_number": {
                    "type": "string",
                    "example": "088999121367"
                },
                "status": {
                    "type": "string",
                    "example": "Activated"
                },
                "sub_area_channel_name": {
                    "type": "string",
                    "example": "0004"
                }
            }
        },
        "models.Meta": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "message": {
                    "type": "string",
                    "example": "OK"
                },
                "status": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "models.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "meta": {
                    "$ref": "#/definitions/models.Meta"
                }
            }
        },
...
        "models.SearchReq": {
            "type": "object",
            "properties": {
                "keyword": {
                    "type": "string",
                    "example": "IDM"
                }
            }
        },
...
    }
}